test: colocate runtime module coverage
